+++ fabric21/src/main/java/com/vecoo/extrartp/ExtraRTP.java
@@ -1,10 +1,12 @@
 package com.vecoo.extrartp;
 
 import com.mojang.logging.LogUtils;
+import com.vecoo.extralib.config.YamlConfigFactory;
 import com.vecoo.extrartp.command.RandomTeleportCommand;
 import com.vecoo.extrartp.config.LocaleConfig;
 import com.vecoo.extrartp.config.ServerConfig;
 import com.vecoo.extrartp.listener.RTPListener;
+import lombok.Getter;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
 import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
@@ -12,13 +14,16 @@ import net.fabricmc.fabric.api.networking.v1.ServerPlayConnectionEvents;
 import net.minecraft.server.MinecraftServer;
 import org.slf4j.Logger;
 
+import java.nio.file.Path;
+
 public class ExtraRTP implements ModInitializer {
     public static final String MOD_ID = "extrartp";
     private static final Logger LOGGER = LogUtils.getLogger();
 
+    @Getter
     private static ExtraRTP instance;
 
-    private ServerConfig config;
+    private ServerConfig serverConfig;
     private LocaleConfig localeConfig;
 
     private MinecraftServer server;
@@ -35,26 +40,16 @@ public class ExtraRTP implements ModInitializer {
     }
 
     public void loadConfig() {
-        try {
-            this.config = new ServerConfig();
-            this.config.init();
-            this.localeConfig = new LocaleConfig();
-            this.localeConfig.init();
-        } catch (Exception e) {
-            LOGGER.error("Error load config.", e);
-        }
-    }
-
-    public static ExtraRTP getInstance() {
-        return instance;
+        this.serverConfig = YamlConfigFactory.load(ServerConfig.class, Path.of("config/ExtraRTP/config.yml"));
+        this.localeConfig = YamlConfigFactory.load(LocaleConfig.class, Path.of("config/ExtraRTP/locale.yml"));
     }
 
     public static Logger getLogger() {
         return LOGGER;
     }
 
-    public ServerConfig getConfig() {
-        return instance.config;
+    public ServerConfig getServerConfig() {
+        return instance.serverConfig;
     }
 
     public LocaleConfig getLocaleConfig() {