+++ fabric21/src/main/java/com/vecoo/extrartp/api/service/ExtraRTPService.java
@@ -1,52 +1,47 @@
-package com.vecoo.extrartp.api.factory;
+package com.vecoo.extrartp.api.service;
 
 import com.vecoo.extrartp.ExtraRTP;
-import com.vecoo.extrartp.api.events.RandomTeleportEvent;
-import com.vecoo.extrartp.config.ServerConfig;
 import com.vecoo.extrartp.util.Utils;
+import lombok.val;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
-import net.minecraft.util.RandomSource;
 import net.minecraft.world.level.block.state.BlockState;
-import net.minecraft.world.level.border.WorldBorder;
 import net.minecraft.world.level.chunk.ChunkAccess;
-import net.minecraft.world.level.chunk.ChunkStatus;
+import net.minecraft.world.level.chunk.status.ChunkStatus;
 import net.minecraft.world.phys.Vec3;
 import org.jetbrains.annotations.NotNull;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.atomic.AtomicBoolean;
 
-public class ExtraRTPFactory {
+public class ExtraRTPService {
     public static boolean randomTeleport(@NotNull ServerPlayer player, @NotNull ServerLevel level) {
-        ServerConfig config = ExtraRTP.getInstance().getConfig();
-        RandomSource random = level.getRandom();
-        WorldBorder worldBorder = level.getWorldBorder();
+        val serverConfig = ExtraRTP.getInstance().getServerConfig();
+        val random = level.getRandom();
+        val worldBorder = level.getWorldBorder();
 
-        double minX = worldBorder.getMinX();
-        double maxX = worldBorder.getMaxX();
-        double minZ = worldBorder.getMinZ();
-        double maxZ = worldBorder.getMaxZ();
-        int y = Utils.heightStart(level.dimension().location().getPath());
+        val minX = worldBorder.getMinX();
+        val maxX = worldBorder.getMaxX();
+        val minZ = worldBorder.getMinZ();
+        val maxZ = worldBorder.getMaxZ();
+        val y = Utils.heightStart(level.dimension().location().getPath());
 
         double deltaX = maxX - minX;
         double deltaZ = maxZ - minZ;
 
-        AtomicBoolean teleportSuccess = new AtomicBoolean(false);
+        val teleportSuccess = new AtomicBoolean(false);
 
-        for (int attempt = 0; attempt < config.getCountAttemptsTeleport() && !teleportSuccess.get(); attempt++) {
-            BlockPos.MutableBlockPos blockPos = new BlockPos.MutableBlockPos(random.nextInt((int) deltaX) + (int) minX, y, random.nextInt((int) deltaZ) + (int) minZ);
-            ChunkAccess chunk = level.getChunkSource().getChunk(blockPos.getX() >> 4, blockPos.getZ() >> 4, ChunkStatus.FEATURES, true);
+        for (int attempt = 0; attempt < serverConfig.getCountAttemptsTeleport() && !teleportSuccess.get(); attempt++) {
+            val blockPos = new BlockPos.MutableBlockPos(random.nextInt((int) deltaX) + (int) minX, y, random.nextInt((int) deltaZ) + (int) minZ);
+            val chunk = level.getChunkSource().getChunk(blockPos.getX() >> 4, blockPos.getZ() >> 4, ChunkStatus.FEATURES, true);
 
             if (chunk != null) {
                 findPosition(blockPos, chunk).thenAccept(success -> {
                     if (success && teleportSuccess.compareAndSet(false, true)) {
                         ExtraRTP.getInstance().getServer().execute(() -> {
-                            RandomTeleportEvent.Successful event = new RandomTeleportEvent.Successful(player, level, blockPos.getX() + 0.5, blockPos.getY() + 1.0, blockPos.getZ() + 0.5, player.getYRot(), player.getXRot());
-
-                            player.teleportTo(event.getLevel(), event.getX(), event.getY(), event.getZ(), event.getYRot(), event.getXRot());
+                            player.teleportTo(level, blockPos.getX() + 0.5, blockPos.getY() + 1.0, blockPos.getZ() + 0.5, player.getYRot(), player.getXRot());
                             player.setDeltaMovement(Vec3.ZERO);
                         });
                     }