+++ fabric21/src/main/java/com/vecoo/extrartp/command/RandomTeleportCommand.java
@@ -6,10 +6,9 @@ import com.vecoo.extralib.chat.UtilChat;
 import com.vecoo.extralib.permission.UtilPermission;
 import com.vecoo.extralib.world.UtilWorld;
 import com.vecoo.extrartp.ExtraRTP;
-import com.vecoo.extrartp.api.factory.ExtraRTPFactory;
-import com.vecoo.extrartp.config.LocaleConfig;
-import com.vecoo.extrartp.config.ServerConfig;
+import com.vecoo.extrartp.api.service.ExtraRTPService;
 import com.vecoo.extrartp.util.Utils;
+import lombok.val;
 import net.minecraft.commands.CommandBuildContext;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.commands.Commands;
@@ -21,7 +20,7 @@ import org.jetbrains.annotations.NotNull;
 public class RandomTeleportCommand {
     public static void register(CommandDispatcher<CommandSourceStack> dispatcher, CommandBuildContext context, Commands.CommandSelection commandSelection) {
         dispatcher.register(Commands.literal("rtp")
-                .requires(p -> UtilPermission.hasPermission(p, "minecraft.command.rtp"))
+                .requires(p -> ExtraRTP.getInstance().getServerConfig().isLowPermission() || UtilPermission.hasPermission(p, "minecraft.command.rtp"))
                 .executes(e -> executeRTP(e.getSource().getPlayerOrException()))
                 .then(Commands.argument("dimension", StringArgumentType.string())
                         .requires(p -> UtilPermission.hasPermission(p, "minecraft.command.rtp.dimension"))
@@ -36,23 +35,25 @@ public class RandomTeleportCommand {
                             return builder.buildFuture();
                         })
                         .executes(s -> executeRTPDimension(s.getSource().getPlayerOrException(), StringArgumentType.getString(s, "dimension")))
+
                         .then(Commands.argument("player", EntityArgument.player())
                                 .requires(s -> UtilPermission.hasPermission(s, "minecraft.command.rtp.dimension.player"))
                                 .executes(e -> executeRTPDimensionPlayer(e.getSource(), StringArgumentType.getString(e, "dimension"), EntityArgument.getPlayer(e, "player")))))
+
                 .then(Commands.literal("reload")
                         .requires(s -> UtilPermission.hasPermission(s, "minecraft.command.rtp.reload"))
                         .executes(e -> executeReload(e.getSource()))));
     }
 
     private static int executeRTP(@NotNull ServerPlayer player) {
-        ServerConfig config = ExtraRTP.getInstance().getConfig();
-        LocaleConfig localeConfig = ExtraRTP.getInstance().getLocaleConfig();
+        val serverConfig = ExtraRTP.getInstance().getServerConfig();
+        val localeConfig = ExtraRTP.getInstance().getLocaleConfig();
 
-        ServerLevel level = UtilWorld.findLevelByName(config.getDefaultWorld());
+        val level = UtilWorld.findLevelByName(serverConfig.getDefaultWorld());
 
         if (level == null) {
             player.sendSystemMessage(UtilChat.formatMessage(localeConfig.getNotDimensionFound()
-                    .replace("%dimension%", config.getDefaultWorld())));
+                    .replace("%dimension%", serverConfig.getDefaultWorld())));
             return 0;
         }
 
@@ -60,11 +61,11 @@ public class RandomTeleportCommand {
             return 0;
         }
 
-        if (ExtraRTPFactory.randomTeleport(player, level)) {
+        if (ExtraRTPService.randomTeleport(player, level)) {
             Utils.COOLDOWN.put(player.getUUID(), System.currentTimeMillis());
 
             player.sendSystemMessage(UtilChat.formatMessage(ExtraRTP.getInstance().getLocaleConfig().getSuccessfulTeleport()
-                    .replace("%dimension%", ExtraRTP.getInstance().getConfig().getDefaultWorld())));
+                    .replace("%dimension%", ExtraRTP.getInstance().getServerConfig().getDefaultWorld())));
         } else {
             player.sendSystemMessage(UtilChat.formatMessage(localeConfig.getFailedTeleport()));
         }
@@ -73,9 +74,8 @@ public class RandomTeleportCommand {
     }
 
     private static int executeRTPDimension(@NotNull ServerPlayer player, @NotNull String dimension) {
-        ServerLevel level = UtilWorld.findLevelByName(dimension);
-
-        LocaleConfig localeConfig = ExtraRTP.getInstance().getLocaleConfig();
+        val localeConfig = ExtraRTP.getInstance().getLocaleConfig();
+        val level = UtilWorld.findLevelByName(dimension);
 
         if (level == null) {
             player.sendSystemMessage(UtilChat.formatMessage(localeConfig.getNotDimensionFound()
@@ -83,9 +83,9 @@ public class RandomTeleportCommand {
             return 0;
         }
 
-        ServerConfig config = ExtraRTP.getInstance().getConfig();
+        val serverConfig = ExtraRTP.getInstance().getServerConfig();
 
-        if (config.isBlacklistWorld() && config.getBlacklistWorldList().contains(dimension.toLowerCase())) {
+        if (serverConfig.isBlacklistWorld() && serverConfig.getBlacklistWorldList().contains(dimension.toLowerCase())) {
             player.sendSystemMessage(UtilChat.formatMessage(localeConfig.getDimensionBlacklist()
                     .replace("%dimension%", dimension)));
             return 0;
@@ -95,7 +95,7 @@ public class RandomTeleportCommand {
             return 0;
         }
 
-        if (ExtraRTPFactory.randomTeleport(player, level)) {
+        if (ExtraRTPService.randomTeleport(player, level)) {
             Utils.COOLDOWN.put(player.getUUID(), System.currentTimeMillis());
 
             player.sendSystemMessage(UtilChat.formatMessage(ExtraRTP.getInstance().getLocaleConfig().getSuccessfulTeleport()
@@ -108,9 +108,8 @@ public class RandomTeleportCommand {
     }
 
     private static int executeRTPDimensionPlayer(@NotNull CommandSourceStack source, @NotNull String dimension, @NotNull ServerPlayer player) {
-        ServerLevel level = UtilWorld.findLevelByName(dimension);
-
-        LocaleConfig localeConfig = ExtraRTP.getInstance().getLocaleConfig();
+        val localeConfig = ExtraRTP.getInstance().getLocaleConfig();
+        val level = UtilWorld.findLevelByName(dimension);
 
         if (level == null) {
             source.sendSystemMessage(UtilChat.formatMessage(localeConfig.getNotDimensionFound()
@@ -118,7 +117,7 @@ public class RandomTeleportCommand {
             return 0;
         }
 
-        if (ExtraRTPFactory.randomTeleport(player, level)) {
+        if (ExtraRTPService.randomTeleport(player, level)) {
             Utils.COOLDOWN.put(player.getUUID(), System.currentTimeMillis());
 
             player.sendSystemMessage(UtilChat.formatMessage(ExtraRTP.getInstance().getLocaleConfig().getSuccessfulTeleport()
@@ -135,9 +134,17 @@ public class RandomTeleportCommand {
     }
 
     private static int executeReload(@NotNull CommandSourceStack source) {
-        ExtraRTP.getInstance().loadConfig();
+        val localeConfig = ExtraRTP.getInstance().getLocaleConfig();
 
-        source.sendSystemMessage(UtilChat.formatMessage(ExtraRTP.getInstance().getLocaleConfig().getConfigReload()));
+        try {
+            ExtraRTP.getInstance().loadConfig();
+        } catch (Exception e) {
+            source.sendSystemMessage(UtilChat.formatMessage(localeConfig.getErrorReload()));
+            ExtraRTP.getLogger().error(e.getMessage());
+            return 0;
+        }
+
+        source.sendSystemMessage(UtilChat.formatMessage(localeConfig.getReload()));
         return 1;
     }
 }