+++ forge20/src/main/java/com/vecoo/extrartp/listener/RTPListener.java
@@ -3,10 +3,8 @@ package com.vecoo.extrartp.listener;
 import com.vecoo.extralib.chat.UtilChat;
 import com.vecoo.extralib.world.UtilWorld;
 import com.vecoo.extrartp.ExtraRTP;
-import com.vecoo.extrartp.api.factory.ExtraRTPFactory;
-import com.vecoo.extrartp.config.LocaleConfig;
-import com.vecoo.extrartp.config.ServerConfig;
-import net.minecraft.server.level.ServerLevel;
+import com.vecoo.extrartp.api.service.ExtraRTPService;
+import lombok.val;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraftforge.common.UsernameCache;
 import net.minecraftforge.event.entity.player.PlayerEvent;
@@ -15,12 +13,12 @@ import net.minecraftforge.eventbus.api.SubscribeEvent;
 public class RTPListener {
     @SubscribeEvent
     public void onPlayerLoggedIn(PlayerEvent.PlayerLoggedInEvent event) {
-        ServerPlayer player = (ServerPlayer) event.getEntity();
-        ServerConfig config = ExtraRTP.getInstance().getConfig();
-        LocaleConfig localeConfig = ExtraRTP.getInstance().getLocaleConfig();
+        val localeConfig = ExtraRTP.getInstance().getLocaleConfig();
+        val player = (ServerPlayer) event.getEntity();
+        val config = ExtraRTP.getInstance().getServerConfig();
 
         if (!UsernameCache.containsUUID(player.getUUID()) && config.isFirstJoinRTP()) {
-            ServerLevel level = UtilWorld.findLevelByName(config.getDefaultWorld());
+            val level = UtilWorld.findLevelByName(config.getDefaultWorld());
 
             if (level == null) {
                 player.sendSystemMessage(UtilChat.formatMessage(localeConfig.getNotDimensionFound()
@@ -28,7 +26,7 @@ public class RTPListener {
                 return;
             }
 
-            if (ExtraRTPFactory.randomTeleport(player, level)) {
+            if (ExtraRTPService.randomTeleport(player, level)) {
                 player.sendSystemMessage(UtilChat.formatMessage(localeConfig.getSuccessfulTeleport()
                         .replace("%dimension%", config.getDefaultWorld())));
             } else {