+++ neoforge21/src/main/java/com/vecoo/extrartp/api/events/RandomTeleportEvent.java
@@ -1,79 +1,23 @@
 package com.vecoo.extrartp.api.events;
 
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 import net.neoforged.bus.api.Event;
 import org.jetbrains.annotations.NotNull;
 
+@Getter
+@Setter
+@AllArgsConstructor
 public abstract class RandomTeleportEvent extends Event {
+    @NotNull
     private final ServerPlayer player;
+    @NotNull
     private ServerLevel level;
     private double x, y, z;
     private float yRot, xRot;
-
-    public RandomTeleportEvent(@NotNull ServerPlayer player, @NotNull ServerLevel level, double x, double y, double z, float yRot, float xRot) {
-        this.player = player;
-        this.level = level;
-        this.x = x;
-        this.y = y;
-        this.z = z;
-        this.yRot = yRot;
-        this.xRot = xRot;
-    }
-
-    @NotNull
-    public ServerPlayer getPlayer() {
-        return this.player;
-    }
-
-    @NotNull
-    public ServerLevel getLevel() {
-        return this.level;
-    }
-
-    public double getX() {
-        return this.x;
-    }
-
-    public double getY() {
-        return this.y;
-    }
-
-    public double getZ() {
-        return this.z;
-    }
-
-    public float getYRot() {
-        return this.yRot;
-    }
-
-    public float getXRot() {
-        return this.xRot;
-    }
-
-    public void setLevel(@NotNull ServerLevel level) {
-        this.level = level;
-    }
-
-    public void setX(double x) {
-        this.x = x;
-    }
-
-    public void setY(double y) {
-        this.y = y;
-    }
-
-    public void setZ(double z) {
-        this.z = z;
-    }
-
-    public void setYRot(float yRot) {
-        this.yRot = yRot;
-    }
-
-    public void setXRot(float xRot) {
-        this.xRot = xRot;
-    }
 
     public static class Successful extends RandomTeleportEvent {
         public Successful(@NotNull ServerPlayer player, @NotNull ServerLevel level, double x, double y, double z, float yRot, float xRot) {