+++ changelog.md
@@ -1,1 +1,4 @@
-* Update depends.
+* Change type config json to yml.
+* Checking reload config and remove reload storage.
+* Added low permission option (Fabric).
+* Improve code.